+++ lib/rapporteur/checker.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Rapporteur
   # The center of the Rapporteur library, Checker manages holding and running
   # the custom checks, holding any application error messages, and provides the
@@ -6,14 +8,10 @@ module Rapporteur
   class Checker
     extend CheckerDeprecations
 
-
     def initialize
-      @messages = MessageList.new(:messages)
-      @errors = MessageList.new(:errors)
       @check_list = CheckList.new
       reset
     end
-
 
     # Public: Add a pre-built or custom check to your status endpoint. These
     # checks are used to test the state of the world of the application, and
@@ -32,13 +30,13 @@ module Rapporteur
     # Returns self.
     # Raises ArgumentError if the given check does not respond to call.
     #
-    def add_check(object_or_nil_with_block=nil, &block)
+    def add_check(object_or_nil_with_block = nil, &block)
       if block_given?
-        @check_list.add(block)
+        check_list.add(block)
       elsif object_or_nil_with_block.respond_to?(:call)
-        @check_list.add(object_or_nil_with_block)
+        check_list.add(object_or_nil_with_block)
       else
-        raise ArgumentError, "A check must respond to #call."
+        raise ArgumentError, 'A check must respond to #call.'
       end
       self
     end
@@ -51,7 +49,7 @@ module Rapporteur
     # Returns self.
     #
     def clear
-      @check_list.clear
+      check_list.clear
       self
     end
 
@@ -76,7 +74,7 @@ module Rapporteur
     #
     def run
       reset
-      @check_list.each do |object|
+      check_list.each do |object|
         object.call(self)
         break if @halted
       end
@@ -116,8 +114,8 @@ module Rapporteur
     #
     # Returns self.
     #
-    def add_error(name, message, i18n_options={})
-      @errors.add(name, message, i18n_options)
+    def add_error(name, message, i18n_options = {})
+      errors.add(name, message, i18n_options)
       self
     end
 
@@ -143,42 +141,46 @@ module Rapporteur
     #
     # Returns self.
     #
-    def add_message(name, message, i18n_options={})
-      @messages.add(name, message, i18n_options)
+    def add_message(name, message, i18n_options = {})
+      messages.add(name, message, i18n_options)
       self
     end
 
     ##
     # Internal: Returns a hash of messages suitable for conversion into JSON.
     #
-    def as_json(args={})
-      @messages.to_hash
+    def as_json(_args = {})
+      messages.to_hash
     end
 
     ##
     # Internal: Used by Rails' JSON serialization to render error messages.
     #
     def errors
-      @errors
+      Thread.current[:rapporteur_errors] ||= MessageList.new(:errors)
     end
 
     ##
     # Internal: Used by Rails' JSON serialization.
     #
     def read_attribute_for_serialization(key)
-      @messages[key]
+      messages[key]
     end
 
     alias read_attribute_for_validation read_attribute_for_serialization
 
-
     private
 
+    attr_reader :check_list
+
+    def messages
+      Thread.current[:rapporteur_messages] ||= MessageList.new(:messages)
+    end
 
     def reset
       @halted = false
-      @messages.clear
-      @errors.clear
+      messages.clear
+      errors.clear
     end
   end
 end