+++ lib/rapporteur/engine.rb
@@ -1,4 +1,7 @@
+# frozen_string_literal: true
+
 module Rapporteur
   class Engine < Rails::Engine
+    isolate_namespace Rapporteur
   end
 end