+++ lib/rapporteur/rspec.rb
@@ -1,43 +1,61 @@
+# frozen_string_literal: true
+
 require 'rapporteur'
 
 shared_examples_for 'a successful status response' do
+  let(:parsed_body) { JSON.parse(response.body) }
+
   it 'responds with HTTP 200' do
-    expect(subject.response_code).to(eq(200))
+    expect(response.response_code).to(eq(200))
   end
 
-  it 'responds with a JSON content header' do
-    expect(subject.content_type).to(eq(Mime[:json]))
-  end
+  context 'the response headers' do
+    it 'contains a Content-Type JSON header' do
+      expect(response.content_type).to(eq(Mime[:json]))
+    end
 
-  it 'responds with valid JSON' do
-    expect { JSON.parse(subject.body) }.not_to(raise_error)
+    it 'does not contain an ETag header' do
+      expect(response.headers).not_to have_key('ETag')
+    end
+
+    it 'contains a Cache-Control header which disables client caching' do
+      expect(response.headers.fetch('Cache-Control')).to eq('no-cache')
+    end
   end
 
   context 'the response payload' do
-    subject { get(status_path) ; JSON.parse(response.body) }
-
     it 'does not contain errors' do
-      expect(subject).not_to(have_key('errors'))
+      expect(parsed_body).not_to(have_key('errors'))
     end
   end
 end
 
 shared_examples_for 'an erred status response' do
+  let(:parsed_body) { JSON.parse(response.body) }
+
   it 'responds with HTTP 500' do
-    expect(subject.response_code).to(eq(500))
+    expect(response.response_code).to(eq(500))
   end
 
-  it 'responds with a JSON content header' do
-    expect(subject.content_type).to(eq(Mime[:json]))
+  context 'the response headers' do
+    it 'contains a Content-Type JSON header' do
+      expect(response.content_type).to(eq(Mime[:json]))
+    end
+
+    it 'does not contain an ETag header' do
+      expect(response.headers).not_to have_key('ETag')
+    end
+
+    it 'contains a Cache-Control header which disables client caching' do
+      expect(response.headers.fetch('Cache-Control')).to eq('no-cache')
+    end
   end
 
-  it 'responds with valid JSON' do
-    expect { JSON.parse(subject.body) }.not_to(raise_error)
-  end
-
-  it 'contains errors' do
-    expect(JSON.parse(subject.body)).to(have_key('errors'))
-    expect(JSON.parse(subject.body).fetch('errors')).not_to(be_empty)
+  context 'the response payload' do
+    it 'contains errors' do
+      expect(parsed_body).to(have_key('errors'))
+      expect(parsed_body.fetch('errors')).not_to(be_empty)
+    end
   end
 end
 
@@ -47,11 +65,11 @@ RSpec::Matchers.define :include_status_error_message do |attribute, message|
     @body.fetch('errors', {}).fetch(attribute.to_s).match(message)
   end
 
-  failure_message_for_should do |actual|
+  failure_message_for_should do |_actual|
     "expected #{@body.inspect} to include a #{attribute}:#{message.inspect} error message"
   end
 
-  failure_message_for_should_not do |actual|
+  failure_message_for_should_not do |_actual|
     "expected #{@body.inspect} to not include a #{attribute}:#{message.inspect} error message"
   end
 end
@@ -59,14 +77,14 @@ end
 RSpec::Matchers.define :include_status_message do |name, message|
   match do |response|
     @body = JSON.parse(response.body)
-    @body.has_key?(name) && @body.fetch(name).match(message)
+    @body.key?(name) && @body.fetch(name).match(message)
   end
 
-  failure_message_for_should do |actual|
+  failure_message_for_should do |_actual|
     "expected #{@body.inspect} to include a #{name.inspect}: #{message.inspect} message"
   end
 
-  failure_message_for_should_not do |actual|
+  failure_message_for_should_not do |_actual|
     "expected #{@body.inspect} to not include a #{name.inspect}: #{message.inspect} message"
   end
 end