+++ spec/requests/no_checks_spec.rb
@@ -1,7 +1,11 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
-describe 'A status request with no checks', :type => :request do
-  subject { get(status_path) ; response }
+RSpec.describe 'A status request with no checks', type: :request do
+  before do
+    get(rapporteur.status_path(format: 'json'))
+  end
 
   it_behaves_like 'a successful status response'
 end